refactor(storage): inline object lock config entrypoints
